Refactor search functionality to use Input component
This commit refactors the search functionality in Searchbar.jsx to use the Input component instead of the Search component from semantic-ui-react. It also adds an useEffect hook that calls a new function, searchMovies(), which fetches data from an API endpoint and passes it to onSearch() prop. The handleSearchChange() function is updated to setQuery using e.target.value instead of {value}.