Add POST request for comments
This commit adds a new POST request to the server.js file that allows users to create comments. The request takes in parameters such as review_id, user_id, comment_text and timestamp and inserts them into the database using SQL queries. If there is an error, it returns a 400 status code with an error message.