Refactor CommentForm component to include user ID in comment object
This commit refactors the CommentForm component to include a user ID in the comment object. The user ID is passed as a prop and stored in state, allowing it to be included when submitting a new comment. This change improves the functionality of the CommentForm component by ensuring that comments are associated with the correct user.