Add delete request for comments
This commit adds a new delete request for comments in the server.js file. The endpoint is "/api/comments/:comment_id". When this endpoint is called, the comment with the specified ID will be deleted from the database. If successful, a 200 status code will be returned and if there's an error, a 400 status code will be returned along with an error message.