Refactor main header and add overlay
This commit refactors the main header by adding an overlay to it. The new overlay is a div with a class of "main-header-overlay" that contains the movie title, release date, and overview. The main header now has a position of relative and height of 500px while the new overlay has a position of absolute, top left corner at (0,0), width and height of 100%, z-index of 1, display flex with justify-content:flex-start; align-items: center; text-align: left; border-radius: 10px.