Refactor movie info display and add backdrop image
This commit refactors the CSS for displaying movie information and adds a backdrop image to the SingleMovie page. The `movieInfoDiv` class was removed, and new classes were added to style the backdrop, poster, and container elements. The `movieInfoContainer` class now positions the movie information on top of the backdrop image.