Refactor SingleMovie component to add CSS classes and improve readability
This commit refactors the SingleMovie component by adding CSS classes to the title and span elements, improving their styling. It also adjusts the width of Grid.Column for cast members' images and adds a centered Grid.Row for the trailer modal. These changes make the code more readable and maintainable.